[DE39] - make get_access_endpoint return search result + me
if 1 node found in search, but I am also a role holder (not found in
search as I am not yet indexed), then return a result set of 2 - myself
and the node found in search.
This will then call get_lb_endpoint and place the appropriate HA IP for
a role into the various config files.