Fixes DE202 - vip bug in get_access_endpoint()
RCB#get_access_endpoint wasn't properly handling the current node when
the node is converging for the first time.  Since the node's run list
hasn't yet been saved to chef server, it doesn't show up in
Chef::Search::Query results.

The existing code only checked this condition when there were:

    (a) ZERO results returned; or
    (b) exactly ONE result returned *and* the result is current node
        (which doesn't even really make sense?)

Now we unconditionally check whether we are a target role-holder, and
if so, add ourselves to the results (unless we're already there).